fix(agent-auth): clear rung-7 repo-shape violations in home onboarding
Repo-shape strict flagged two rung-7 violations:
- RAW_DOM_CONTROL: raw <button> affordance in the auth-setup evidence row. Swap
  for the Button primitive (variant/size unstyled), preserving the inline
  underline styling.
- LARGE_FRONTEND_FILE: HomeOnboardingCards at 419 lines. Extract the evidence
  card cluster (formatProbeCountdown + AuthSetupEvidenceRow + AuthSetupEvidenceCard)
  to a sibling HomeOnboardingEvidenceCard.tsx; the file drops to 273 lines.

Structure report strict: TOTAL 0. home vitest 61 passed, tsc clean.
